+++ xrpl/models/transactions/loan_broker_cover_withdraw.py
@@ -32,6 +32,13 @@ class LoanBrokerCoverWithdraw(Transaction):
     An account to receive the assets. It must be able to receive the asset.
     """
 
+    destination_tag: Optional[int] = None
+    """
+    An arbitrary `destination tag
+    <https://xrpl.org/source-and-destination-tags.html>`_ that
+    identifies the reason for the Payment, or a hosted recipient to pay.
+    """
+
     transaction_type: TransactionType = field(
         default=TransactionType.LOAN_BROKER_COVER_WITHDRAW,
         init=False,